requirements.txt
pandas==2.2.3
seaborn==0.13.2
numpy==2.2.1
matplotlib-inline==0.1.7
streamlit==1.41.1
xlrd @ file:///home/conda/feedstock_root/build_artifacts/xlrd_1610224409810/work
XlsxWriter==3.2.0
protobuf==3.20.1